test: relocate #6032 ollama health tests to part_03 (file-layout limit)
The four new health-status/degraded-diagnostics tests pushed
ollama_admin_tests_part_01_tests.rs to 761 lines, over the repo's 750-line
per-file layout limit (Enforce OpenHuman Rust file layout gate). Moved them to
the sibling part_03 file (386 → 510 lines); no test logic changed.
